+++ components/ClientHandlerBar/styles.js
@@ -1,20 +1,23 @@
 import styled from "styled-components";
-import { BORDER_RADIUS, COLORS } from "../../common/styles/styles";
+import { BORDER_RADIUS, COLORS, BREAKPOINTS } from "../../common/styles/styles";
 
 export const Container = styled.div`
   width: 100%;
   display: flex;
+  overflow-x: auto;
 `;
 
 export const Text = styled.p`
-  padding: 5px 10px;
+  padding: 5px 0;
   color: ${COLORS.primary};
   font-size: 16px;
   user-select: none;
+  font-family: "Rubik-Medium";
 `;
 
 export const List = styled.div`
   display: flex;
+  flex-wrap: nowrap;
 `;
 
 export const Item = styled.p`
@@ -28,11 +31,11 @@ export const Item = styled.p`
   user-select: none;
   transition: 0.3s;
 
-  &:first-child {
-    margin-left: 0;
-  }
-
   &:hover {
     background-color: ${COLORS.white};
+  }
+
+  &:last-child {
+    margin-right: 16px;
   }
 `;
